--- src/is-initializer.ts
@@ -1,5 +1,0 @@
-export function isInitializer<V>(
-  initialValue: V | (() => V),
-): initialValue is () => V {
-  return typeof initialValue === 'function';
-}
