+++ src/utils/batched-updates.ts
