+++ src/utils/type-utils.ts
@@ -1,5 +1,8 @@
 export type Defined<T> = T extends undefined ? never : T;
 
+export const isDefined = <V>(value: V | undefined): value is Defined<V> =>
+  value !== undefined;
+
 export declare type IsNever<T> = [T] extends [never] ? true : false;
 export type NonNever<T> = IsNever<T> extends true ? any : T;
 
@@ -9,7 +12,7 @@ export type MethodKeys<T> = {
 export type Methods<T> = Pick<T, MethodKeys<T>>;
 export type KeyOfMethods<T> = NonNever<keyof Methods<T>>;
 export type CB<T> = (t: T) => void;
-export type StrictGuard<T extends (...args: any) => any> = [T, CB<T>];
-export type StrictMethodsGuard<Fs> = {
-  [P in keyof Methods<Fs>]: StrictGuard<Fs[P]>;
+export type StrictGuard<T> = [T, CB<T>];
+export type StrictMethodsGuard<Fns> = {
+  [P in keyof Methods<Fns>]: StrictGuard<Fns[P]>;
 };
